requirements.txt
mediapipe
pywebview
